requirements.txt
kaggle
matplotlib
numpy
pandas
Pillow
python-dotenv
scikit-image
scikit-learn
seaborn
tqdm
transformers
